% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-package.R
\docType{package}
\name{betaMC-package}
\alias{betaMC-package}
\alias{_PACKAGE}
\title{betaMC: Monte Carlo Confidence Intervals for Standardized Regression Coefficients}
\description{
Generates Monte Carlo confidence intervals for standardized regression coefficients for models fitted by lm(). 'betaMC' combines ideas from Monte Carlo confidence intervals for the indirect effect (Preacher and Selig, 2012 \doi{10.1080/19312458.2012.679848}) and the sampling covariance matrix of regression coefficients (Dudgeon, 2017 \doi{10.1007/s11336-017-9563-z}) to generate confidence intervals for standardized regression coefficients.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeksterslab/betaMC}
  \item \url{https://jeksterslab.github.io/betaMC/}
  \item Report bugs at \url{https://github.com/jeksterslab/betaMC/issues}
}

}
\author{
\strong{Maintainer}: Ivan Jacob Agaloos Pesigan \email{r.jeksterslab@gmail.com} (\href{https://orcid.org/0000-0003-4818-8420}{ORCID}) [copyright holder]

}
\keyword{internal}
