% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-beta-mc.R
\name{coef.betamc}
\alias{coef.betamc}
\title{Standardized Regression Slopes}
\usage{
\method{coef}{betamc}(object, ...)
}
\arguments{
\item{object}{Object of class \code{betamc}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of standardized regression slopes.
}
\description{
Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object, R = 100)
# use a large R, for example, R = 20000 for actual research
coef(std)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
