% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-diff-beta-mc.R
\name{coef.diffbetamc}
\alias{coef.diffbetamc}
\title{Differences of Standardized Regression Slopes}
\usage{
\method{coef}{diffbetamc}(object, ...)
}
\arguments{
\item{object}{Object of class \code{diffbetamc}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of differences of standardized regression slopes.
}
\description{
Differences of Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object, R = 100)
# use a large R, for example, R = 20000 for actual research
diff <- DiffBetaMC(std)
coef(diff)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
