% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-p-cor-beta-mc.R
\name{coef.pcorbetamc}
\alias{coef.pcorbetamc}
\title{Partial and Semipartial Correlation Coefficients}
\usage{
\method{coef}{pcorbetamc}(object, ...)
}
\arguments{
\item{object}{Object of class \code{pcorbetamc}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of partial and semipartial correlation coefficients
}
\description{
Partial and Semipartial Correlation Coefficients
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object, R = 100)
# use a large R, for example, R = 20000 for actual research
pcor <- PCorBetaMC(std)
coef(pcor)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
