% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-r-sq-beta-mc.R
\name{coef.rsqbetamc}
\alias{coef.rsqbetamc}
\title{Multiple Correlation Coefficients
(R-Squared and Adjusted R-Squared)}
\usage{
\method{coef}{rsqbetamc}(object, ...)
}
\arguments{
\item{object}{Object of class \code{rsqbetamc}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of multiple correlation coefficients
(R-squared and adjusted R-squared)
}
\description{
Multiple Correlation Coefficients
(R-Squared and Adjusted R-Squared)
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object, R = 100)
# use a large R, for example, R = 20000 for actual research
rsq <- RSqBetaMC(std)
coef(rsq)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
