% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-r-sq-beta-mc.R
\name{confint.rsqbetamc}
\alias{confint.rsqbetamc}
\title{Confidence Intervals for Differences
of Standardized Regression Slopes}
\usage{
\method{confint}{rsqbetamc}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Object of class \code{rsqbetamc}.}

\item{parm}{a specification of which parameters
are to be given confidence intervals,
either a vector of numbers or a vector of names.
If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of confidence intervals.
}
\description{
Confidence Intervals for Differences
of Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object)
rsq <- RSqBetaMC(std)
confint(rsq, level = 0.95)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
