% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-generic-dif.R
\name{dif}
\alias{dif}
\title{Differences of Regression Slopes}
\usage{
dif(object, ...)
}
\arguments{
\item{object}{Object used to select a method.}

\item{...}{additional arguments.}
}
\description{
Differences of Regression Slopes
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
