% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-mc-mi.R
\name{MCMI}
\alias{MCMI}
\title{Generate the Sampling Distribution of Regression Parameters
Using the Monte Carlo Method for Data with Missing Values}
\usage{
MCMI(
  object,
  R = 20000L,
  type = "hc3",
  g1 = 1,
  g2 = 1.5,
  k = 0.7,
  decomposition = "eigen",
  pd = TRUE,
  tol = 1e-06,
  fixed_x = FALSE,
  seed_mc = NULL,
  adj = FALSE,
  seed_mi = NA,
  fun = "mice",
  imp = NULL,
  ...
)
}
\arguments{
\item{object}{Object of class \code{lm}.}

\item{R}{Positive integer.
Number of Monte Carlo replications.}

\item{type}{Character string.
Sampling covariance matrix type.
Possible values are
\code{"mvn"},
\code{"adf"},
\code{"hc0"},
\code{"hc1"},
\code{"hc2"},
\code{"hc3"},
\code{"hc4"},
\code{"hc4m"}, and
\code{"hc5"}.
\code{type = "mvn"} uses the normal-theory sampling covariance matrix.
\code{type = "adf"} uses the asymptotic distribution-free
sampling covariance matrix.
\code{type = "hc0"} through \code{"hc5"} uses different versions of
heteroskedasticity-consistent sampling covariance matrix.}

\item{g1}{Numeric.
\code{g1} value for \code{type = "hc4m"} or \code{type = "hc5"}.}

\item{g2}{Numeric.
\code{g2} value for \code{type = "hc4m"}.}

\item{k}{Numeric.
Constant for \code{type = "hc5"}}

\item{decomposition}{Character string.
Matrix decomposition of the sampling variance-covariance matrix
for the data generation.
If \code{decomposition = "chol"}, use Cholesky decomposition.
If \code{decomposition = "eigen"}, use eigenvalue decomposition.
If \code{decomposition = "svd"}, use singular value decomposition.}

\item{pd}{Logical.
If \code{pd = TRUE},
check if the sampling variance-covariance matrix
is positive definite using \code{tol}.}

\item{tol}{Numeric.
Tolerance used for \code{pd}.}

\item{fixed_x}{Logical.
If \code{fixed_x = TRUE}, treat the regressors as fixed.
If \code{fixed_x = FALSE}, treat the regressors as random.}

\item{seed_mc}{Integer.
Random seed for the Monte Carlo method.}

\item{adj}{Logical.
If \code{adj = TRUE},
use Li, Raghunathan, and Rubin (1991)
sampling covariance matrix adjustment.
If \code{adj = FALSE},
use the multivariate version of Rubin's (1987)
sampling covariance matrix.}

\item{seed_mi}{Integer.
Random seed for multiple imputation.}

\item{fun}{Character string.
Multiple imputation function.
If \code{fun = "mice"}, use \code{\link[mice:mice]{mice::mice()}}.
If \code{fun = "amelia"}, use \code{\link[Amelia:amelia]{Amelia::amelia()}}.}

\item{imp}{Optional argument.
A list of multiply imputed data sets.}

\item{...}{Additional arguments to pass to \code{fun}.
If \code{fun = "mice"}, DO NOT supply \code{data}, \code{seed}, or \code{print}.
If \code{fun = "amelia"}, DO NOT supply \code{x} or \code{p2s}.}
}
\value{
Returns an object
of class \code{mc} which is a list with the following elements:
\describe{
\item{call}{Function call.}
\item{args}{Function arguments.}
\item{lm_process}{Processed \code{lm} object.}
\item{scale}{Sampling variance-covariance matrix of parameter estimates.}
\item{location}{Parameter estimates.}
\item{thetahatstar}{Sampling distribution of parameter estimates.}
\item{fun}{Function used ("MCMI").}
}
}
\description{
Generate the Sampling Distribution of Regression Parameters
Using the Monte Carlo Method for Data with Missing Values
}
\details{
Multiple imputation (\code{mice::mice()})
is used to deal with missing values in a data set.
The vector of parameter estimates
and the corresponding sampling covariance matrix
are estimated for each of the imputed data sets.
Results are combined to arrive at the pooled vector of parameter estimates
and the corresponding sampling covariance matrix.
The pooled estimates are then used to generate the sampling distribution
of regression parameters.
See \code{MC()} for more details on the Monte Carlo method.
}
\examples{
set.seed(42)
nas1982_missing <- mice::ampute(nas1982)$amp
# Fit the regression model
## Note that this does not deal with missing values.
## The fitted model (`object`) is updated with each imputed data
## within the `MCMI()` function.
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982_missing)
# Generate the sampling distribution of parameter estimates
# (use large values R and m, for example, R = 20000 and m = 100,
# for actual research)
MCMI(object, R = 100, M = 5)
}
\references{
Dudgeon, P. (2017).
Some improvements in confidence intervals
for standardized regression coefficients.
\emph{Psychometrika}, \emph{82}(4), 928–951.
\doi{10.1007/s11336-017-9563-z}

Pesigan, I. J. A., & Cheung, S. F. (2023).
Monte Carlo confidence intervals for the indirect effect with missing data.
\emph{Behavior Research Methods}.
\doi{10.3758/s13428-023-02114-4}

Preacher, K. J., & Selig, J. P. (2012).
Advantages of Monte Carlo confidence intervals for indirect effects.
\emph{Communication Methods and Measures}, \emph{6}(2), 77-98.
\doi{10.1080/19312458.2012.679848}
}
\seealso{
Other Beta Monte Carlo Functions: 
\code{\link{BetaMC}()},
\code{\link{DeltaRSqMC}()},
\code{\link{DiffBetaMC}()},
\code{\link{MC}()},
\code{\link{PCorMC}()},
\code{\link{RSqMC}()},
\code{\link{SCorMC}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Monte Carlo Functions}
\keyword{betaMC}
\keyword{mc}
