% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSets-nas1982.R
\docType{data}
\name{nas1982}
\alias{nas1982}
\title{1982 National Academy of Sciences Doctoral Programs Data}
\format{
Ratings of 46 doctoral programs in psychology in the USA
with the following variables:
\describe{
\item{QUALITY}{Program quality ratings.}
\item{NFACUL}{Number of faculty members in the program.}
\item{NGRADS}{Number of program graduates.}
\item{PCTSUPP}{Percentage of program graduates who received support.}
\item{PCTGRT}{Percent of faculty members holding research grants.}
\item{NARTIC}{Number of published articles
attributed to program faculty member.}
\item{PCTPUB}{Percent of faculty with one or more published article.}
}
}
\usage{
nas1982
}
\description{
1982 National Academy of Sciences Doctoral Programs Data
}
\references{
National Research Council. (1982).
\emph{An assessment of research-doctorate programs in the United States:
Social and behavioral sciences}.
\doi{10.17226/9781}.
Reproduced with permission from the National Academy of Sciences,
Courtesy of the National Academies Press, Washington, D.C.
}
\keyword{data}
