% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-p-cor-beta-mc.R
\name{PCorBetaMC}
\alias{PCorBetaMC}
\title{Estimate Partial and Semipartial Correlation Coefficients
and Generate the Corresponding Sampling Distribution
Using the Monte Carlo Method}
\usage{
PCorBetaMC(object)
}
\arguments{
\item{object}{Object of class \code{betamc},
that is,
the output of the \code{BetaMC()} function.}
}
\value{
Returns an object of class \code{rsqbetamc}
which is a list with the following elements:
\describe{
\item{fit}{The argument \code{object}.}
\item{thetahatstar}{Sampling distribution of
partial and semipartial correlation coefficients.}
\item{vcov}{Sampling covariance matrix of
partial and semipartial correlation coefficients.}
\item{est}{Vector of
partial and semipartial correlation coefficients.}
}
}
\description{
Estimate Partial and Semipartial Correlation Coefficients
and Generate the Corresponding Sampling Distribution
Using the Monte Carlo Method
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object, R = 100)
# use a large R, for example, R = 20000 for actual research
pcor <- PCorBetaMC(std)
# Methods -------------------------------------------------------
print(pcor)
summary(pcor)
coef(pcor)
vcov(pcor)
confint(pcor, level = 0.95)
}
\seealso{
Other Beta Monte Carlo Functions: 
\code{\link{BetaMC}()},
\code{\link{DiffBetaMC}()},
\code{\link{RSqBetaMC}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Monte Carlo Functions}
\keyword{betaMC}
\keyword{pcor}
