% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-beta-mc.R
\name{print.betamc}
\alias{print.betamc}
\title{Print Method for an Object of Class
\code{betamc}}
\usage{
\method{print}{betamc}(x, alpha = NULL, digits = 4, ...)
}
\arguments{
\item{x}{Object of Class \code{betamc},
that is,
the output of the
\code{BetaMC()},
\code{RSqMC()},
\code{SCorMC()},
\code{DeltaRSqMC()},
\code{PCorMC()}, or
\code{DiffBetaMC()}
functions.}

\item{alpha}{Numeric vector.
Significance level \eqn{\alpha}.
If \code{alpha = NULL},
use the argument \code{alpha} used in \code{x}.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Prints a matrix of
estimates,
standard errors,
number of Monte Carlo replications,
and
confidence intervals.
}
\description{
Print Method for an Object of Class
\code{betamc}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
