% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-mc.R
\name{print.mc}
\alias{print.mc}
\title{Print Method for an Object of Class \code{mc}}
\usage{
\method{print}{mc}(x, ...)
}
\arguments{
\item{x}{Object of Class \code{mc}.}

\item{...}{additional arguments.}
}
\value{
Prints the first set of simulated parameter estimates
and model-implied covariance matrix.
}
\description{
Print Method for an Object of Class \code{mc}
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
mc <- MC(object, R = 100)
print(mc)

}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
