% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-r-sq-beta-mc.R
\name{print.rsqbetamc}
\alias{print.rsqbetamc}
\title{Print Method for an Object of Class \code{rsqbetamc}}
\usage{
\method{print}{rsqbetamc}(x, alpha = c(0.05, 0.01, 0.001), digits = 4, ...)
}
\arguments{
\item{x}{Object of class \code{rsqbetamc}.}

\item{alpha}{Significance level.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of
multiple correlation
(R-squared and adjusted R-squared),
standard errors,
number of Monte Carlo replications,
and
confidence intervals.
}
\description{
Print Method for an Object of Class \code{rsqbetamc}
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object, R = 100)
# use a large R, for example, R = 20000 for actual research
rsq <- RSqBetaMC(std)
print(rsq)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
