% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-generic-rsq.R,
%   R/betaMC-methods-beta-mc-r-sq-beta-mc.R
\name{rsq}
\alias{rsq}
\alias{rsq.betamc}
\title{Estimate Coefficient of Determination
(R-Squared and Adjusted R-Squared)
and Sampling Covariance Matrix}
\usage{
rsq(object, ...)

\method{rsq}{betamc}(object, ...)
}
\arguments{
\item{object}{Object used to select a method.}

\item{...}{additional arguments.}
}
\description{
Estimate Coefficient of Determination
(R-Squared and Adjusted R-Squared)
and Sampling Covariance Matrix
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
