% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-r-sq-beta-mc.R
\name{RSqBetaMC}
\alias{RSqBetaMC}
\title{Estimate Differences of Standardized Slopes
and Sampling Covariance Matrix}
\usage{
RSqBetaMC(object)
}
\arguments{
\item{object}{Object of class \code{betamc},
that is,
the output of the \code{BetaMC()} function.}
}
\value{
Returns an object of class \code{rsqbetamc}
which is a list with the following elements:
\describe{
\item{fit}{The argument \code{object}.}
\item{thetahatstar}{Sampling distribution.}
\item{vcov}{Sampling covariance matrix of
differences of standardized slopes.}
\item{est}{Vector of
differences of standardized slopes.}
}
}
\description{
Estimate Differences of Standardized Slopes
and Sampling Covariance Matrix
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object)
rsq <- RSqBetaMC(std)
# Methods -------------------------------------------------------
print(rsq)
summary(rsq)
coef(rsq)
vcov(rsq)
confint(rsq, level = 0.95)
}
\seealso{
Other Beta Monte Carlo Functions: 
\code{\link{BetaMC}()},
\code{\link{DiffBetaMC}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Monte Carlo Functions}
\keyword{betaMC}
\keyword{rsq}
