% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-p-cor-beta-mc.R
\name{summary.pcorbetamc}
\alias{summary.pcorbetamc}
\title{Summary Method for an Object of Class \code{pcorbetamc}}
\usage{
\method{summary}{pcorbetamc}(object, alpha = c(0.05, 0.01, 0.001), digits = 4, ...)
}
\arguments{
\item{object}{Object of class \code{pcorbetamc}.}

\item{alpha}{Significance level.}

\item{digits}{Digits to print.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of
partial and semipartial correlation coefficients,
standard errors,
number of Monte Carlo replications,
and
confidence intervals.
}
\description{
Summary Method for an Object of Class \code{pcorbetamc}
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object, R = 100)
# use a large R, for example, R = 20000 for actual research
pcor <- PCorBetaMC(std)
summary(pcor)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
