% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-beta-mc.R
\name{vcov.betamc}
\alias{vcov.betamc}
\title{Sampling Variance-Covariance Matrix Method for an Object of Class
\code{betamc}}
\usage{
\method{vcov}{betamc}(object, ...)
}
\arguments{
\item{object}{Object of Class \code{betamc},
that is,
the output of the
\code{BetaMC()},
\code{RSqMC()},
\code{SCorMC()},
\code{DeltaRSqMC()},
\code{PCorMC()}, or
\code{DiffBetaMC()}
functions.}

\item{...}{additional arguments.}
}
\value{
Returns the variance-covariance matrix of estimates.
}
\description{
Sampling Variance-Covariance Matrix Method for an Object of Class
\code{betamc}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
