% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-p-cor-beta-mc.R
\name{vcov.pcorbetamc}
\alias{vcov.pcorbetamc}
\title{Sampling Covariance Matrix of
Partial and Semipartial Correlation Coefficients}
\usage{
\method{vcov}{pcorbetamc}(object, ...)
}
\arguments{
\item{object}{Object of class \code{pcorbetamc}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of the
variance-covariance matrix
of partial and semipartial correlation coefficients.
}
\description{
Sampling Covariance Matrix of
Partial and Semipartial Correlation Coefficients
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object, R = 100)
# use a large R, for example, R = 20000 for actual research
pcor <- PCorBetaMC(std)
vcov(pcor)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
