% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-r-sq-beta-mc.R
\name{vcov.rsqbetamc}
\alias{vcov.rsqbetamc}
\title{Sampling Covariance Matrix of
Differences of Standardized Regression Slopes}
\usage{
\method{vcov}{rsqbetamc}(object, ...)
}
\arguments{
\item{object}{Object of class \code{rsqbetamc}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of the variance-covariance matrix
of differences of standardized regression slopes.
}
\description{
Sampling Covariance Matrix of
Differences of Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object)
out <- rsq(std)
vcov(out)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
