% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-beta-mc.R
\name{coef.betamc}
\alias{coef.betamc}
\title{Estimated Parameter Method for an Object of Class
\code{betamc}}
\usage{
\method{coef}{betamc}(object, ...)
}
\arguments{
\item{object}{Object of Class \code{betamc},
that is,
the output of the
\code{BetaMC()},
\code{RSqMC()},
\code{SCorMC()},
\code{DeltaRSqMC()},
\code{PCorMC()}, or
\code{DiffBetaMC()}
functions.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of estimated parameters.
}
\description{
Estimated Parameter Method for an Object of Class
\code{betamc}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
