% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-r-sq-beta-mc.R
\name{coef.rsqbetamc}
\alias{coef.rsqbetamc}
\title{Differences of Standardized Regression Slopes}
\usage{
\method{coef}{rsqbetamc}(object, ...)
}
\arguments{
\item{object}{Object of class \code{rsqbetamc}.}

\item{...}{additional arguments.}
}
\value{
Returns a vector of multiple correlation.
}
\description{
Differences of Standardized Regression Slopes
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object)
rsq <- RSqBetaMC(std)
coef(rsq)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
