% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/001-generic-dif.R,
%   R/betaMC-methods-beta-mc-diff-beta-mc.R
\name{dif}
\alias{dif}
\alias{dif.betamc}
\title{Differences of Regression Slopes}
\usage{
dif(object, ...)

\method{dif}{betamc}(object, ...)
}
\arguments{
\item{object}{Object used to select a method.}

\item{...}{additional arguments.}
}
\description{
Differences of Regression Slopes
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
