% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-diff-beta-mc.R
\name{DiffBetaMC}
\alias{DiffBetaMC}
\title{Estimate Differences of Standardized Slopes
and Generate the Corresponding Sampling Distribution
Using the Monte Carlo Method}
\usage{
DiffBetaMC(object)
}
\arguments{
\item{object}{Object of class \code{betamc},
that is,
the output of the \code{BetaMC()} function.}
}
\value{
Returns an object of class \code{diffbetamc}
which is a list with the following elements:
\describe{
\item{fit}{The argument \code{object}.}
\item{thetahatstar}{Sampling distribution of
differences of standardized slopes.}
\item{vcov}{Sampling covariance matrix of
differences of standardized slopes.}
\item{est}{Vector of
differences of standardized slopes.}
}
}
\description{
Estimate Differences of Standardized Slopes
and Generate the Corresponding Sampling Distribution
Using the Monte Carlo Method
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object, R = 100)
# use a large R, for example, R = 20000 for actual research
diff <- DiffBetaMC(std)
# Methods -------------------------------------------------------
print(diff)
summary(diff)
coef(diff)
vcov(diff)
confint(diff, level = 0.95)
}
\seealso{
Other Beta Monte Carlo Functions: 
\code{\link{BetaMC}()},
\code{\link{PCorBetaMC}()},
\code{\link{RSqBetaMC}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Monte Carlo Functions}
\keyword{betaMC}
\keyword{diff}
