% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-diff-beta-mc.R
\name{DiffBetaMC}
\alias{DiffBetaMC}
\title{Estimate Differences of Standardized Slopes
and Generate the Corresponding Sampling Distribution
Using the Monte Carlo Method}
\usage{
DiffBetaMC(object)
}
\arguments{
\item{object}{Object of class \code{mc}, that is,
the output of the \code{MC()} function.}
}
\value{
Returns an object
of class \code{betamc} which is a list with the following elements:
\describe{
\item{object}{The function argument \code{object}.}
\item{thetahatstar}{Sampling distribution of
differences of standardized regression slopes.}
\item{vcov}{Sampling variance-covariance matrix of
differences of standardized regression slopes.}
\item{est}{Vector of estimated
differences of standardized regression slopes.}
\item{fun}{Function used ("DiffBetaMC").}
}
}
\description{
Estimate Differences of Standardized Slopes
and Generate the Corresponding Sampling Distribution
Using the Monte Carlo Method
}
\details{
The vector of differences of standardized regression slopes
is derived from each randomly generated vector of parameter estimates.
Confidence intervals are generated by obtaining
percentiles corresponding to \eqn{100(1 - \alpha)\%}
from the generated sampling
distribution of differences of standardized regression slopes,
where \eqn{\alpha} is the significance level.
}
\examples{
# Fit the regression model
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
# Generate the sampling distribution of parameter estimates
# (use a large R, for example, R = 20000 for actual research)
mc <- MC(object, R = 100)
# Generate confidence intervals
# for differences of standardized regression slopes
diff <- DiffBetaMC(mc)
# Methods --------------------------------------------------------
print(diff)
summary(diff)
coef(diff)
vcov(diff)
confint(diff, level = 0.95)
}
\seealso{
Other Beta Monte Carlo Functions: 
\code{\link{BetaMC}()},
\code{\link{DeltaRSqMC}()},
\code{\link{MC}()},
\code{\link{PCorMC}()},
\code{\link{RSqMC}()},
\code{\link{SCorMC}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Monte Carlo Functions}
\keyword{betaMC}
\keyword{diff}
