% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-r-sq-beta-mc.R
\name{RSqBetaMC}
\alias{RSqBetaMC}
\title{Estimate Multiple Correlation Coefficients
(R-Squared and Adjusted R-Squared)
and Generate the Corresponding Sampling Distribution
Using the Monte Carlo Method}
\usage{
RSqBetaMC(object)
}
\arguments{
\item{object}{Object of class \code{betamc},
that is,
the output of the \code{BetaMC()} function.}
}
\value{
Returns an object
of class \code{rsqbetamc} which is a list with the following elements:
\describe{
\item{fit}{The argument \code{object}.}
\item{thetahatstar}{Sampling distribution of
multiple correlation coefficients
(R-squared and adjusted R-squared).}
\item{vcov}{Sampling covariance matrix of
multiple correlation coefficients
(R-squared and adjusted R-squared).}
\item{est}{Vector of
multiple correlation coefficients
(R-squared and adjusted R-squared).}
}
}
\description{
Estimate Multiple Correlation Coefficients
(R-Squared and Adjusted R-Squared)
and Generate the Corresponding Sampling Distribution
Using the Monte Carlo Method
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object, R = 100)
# use a large R, for example, R = 20000 for actual research
rsq <- RSqBetaMC(std)
# Methods -------------------------------------------------------
print(rsq)
summary(rsq)
coef(rsq)
vcov(rsq)
confint(rsq, level = 0.95)
}
\seealso{
Other Beta Monte Carlo Functions: 
\code{\link{BetaMC}()},
\code{\link{DiffBetaMC}()},
\code{\link{PCorBetaMC}()}
}
\author{
Ivan Jacob Agaloos Pesigan
}
\concept{Beta Monte Carlo Functions}
\keyword{betaMC}
\keyword{rsq}
