% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaMC-methods-r-sq-beta-mc.R
\name{vcov.rsqbetamc}
\alias{vcov.rsqbetamc}
\title{Sampling Covariance Matrix of
Multiple Correlation Coefficients
(R-Squared and Adjusted R-Squared)}
\usage{
\method{vcov}{rsqbetamc}(object, ...)
}
\arguments{
\item{object}{Object of class \code{rsqbetamc}.}

\item{...}{additional arguments.}
}
\value{
Returns a matrix of the
variance-covariance matrix
of multiple correlation coefficients
(R-squared and adjusted R-squared).
}
\description{
Sampling Covariance Matrix of
Multiple Correlation Coefficients
(R-Squared and Adjusted R-Squared)
}
\examples{
object <- lm(QUALITY ~ NARTIC + PCTGRT + PCTSUPP, data = nas1982)
std <- BetaMC(object, R = 100)
# use a large R, for example, R = 20000 for actual research
rsq <- RSqBetaMC(std)
vcov(rsq)
}
\author{
Ivan Jacob Agaloos Pesigan
}
\keyword{methods}
